refactor(document_extractor): update module structure
- Renamed `models.py` to `entities.py` and updated imports accordingly for better clarity and consistency.
- Renamed `document_extractor_node.py` to `node.py` to simplify module naming.
- Adjusted related imports in test files to reflect the new file names.
