+++ api/core/workflow/nodes/document_extractor/entities.py
