+++ api/tests/unit_tests/core/workflow/nodes/test_document_extractor_node.py
@@ -7,7 +7,7 @@ from core.variables import ArrayFileSegment
 from core.variables.variables import StringVariable
 from core.workflow.entities.node_entities import NodeRunResult
 from core.workflow.nodes.document_extractor import DocumentExtractorNode, DocumentExtractorNodeData
-from core.workflow.nodes.document_extractor.document_extractor_node import (
+from core.workflow.nodes.document_extractor.node import (
     _extract_text_from_doc,
     _extract_text_from_pdf,
     _extract_text_from_plain_text,
